auth/requirements.txt
flask
flask-sqlalchemy
pyjwt
requests